move DefaultAbstractUserawareCustomer functionality to trait
